Rename /Navbar.tsx to /Navbar/index.tsx
